Huffman decoder optimization: to check if code is too long for the lookup table, set nb to HUFF_LOOKAHEAD + 1 explicitely
The compiler then can optimize the upcoming expressions better (apply a
constant mask and the address compuration involving nb)

For reasons that I do not understand, on some platforms (in particular
AMD Phenom II), this is also faster then checking that nb == HUFF_LOOKAHEAD + 1,
and not set nb explicitely afterwards. This is even though one instruction
is generated less then, and the compiler can do the same optimizations.