Huffman decoder optimization: swapped symbol and number of bits in the lookup table
This seems to be faster. Probably this is because the check for a long
code with that change depends on a simpler operation - nb is calculated by
masking instead of shifting. The shift required to get the symbol is
hidden by the code to fill the bit buffer afterwards, that a super-scalar
CPU probably processes in parallel.